Shimming with bundling dependencies inside JARs
This is a dead end because it would require to detect shimming versions when copying packages which would lead to the need to rewrite package.json files adding the shims and so on. Also even if it could be implemented you could virtually have several versions of shimmed node.js’ running in the same page, which could lead to inconsistencies, given that node.js is versioned as a whole, not as a graph of packages.